renamed package 'data' into 'model'
deleted CsarArtifact class
