--- core/src/main/java/org/opentosca/toscana/core/data/CsarArtifact.java
@@ -1,6 +1,0 @@
-package org.opentosca.toscana.core.data;
-
-public class CsarArtifact {
-    // TODO
-    // class representig a csar
-}
