+++ core/src/main/java/org/opentosca/toscana/core/model/Csar.java
@@ -1,4 +1,4 @@
-package org.opentosca.toscana.core.data;
+package org.opentosca.toscana.core.model;
 
 import java.util.Map;
 
@@ -9,13 +9,13 @@ public class Csar {
      * Key is the platform identifier.
      */
     private Map<String, Transformation> transformations;
-    private CsarArtifact artifact;
+    private String artifactPath;
 
-    public Csar(CsarArtifact artifact){
-        this.artifact = artifact;
+    public Csar(String csarArtifactPath) {
+        this.artifactPath = csarArtifactPath;
     }
 
-    public void transform(Platform targetPlatform){
+    public void transform(Platform targetPlatform) {
         // TODO
         throw new UnsupportedOperationException();
     }
