+++ core/src/main/java/org/opentosca/toscana/core/model/LogLevel.java
@@ -1,4 +1,4 @@
-package org.opentosca.toscana.core.data;
+package org.opentosca.toscana.core.model;
 
 public enum LogLevel {
     TRACE, DEBUG, INFO, WARNING, ERROR
