+++ core/src/main/java/org/opentosca/toscana/core/model/TransformationState.java
@@ -1,4 +1,4 @@
-package org.opentosca.toscana.core.data;
+package org.opentosca.toscana.core.model;
 
 /**
  * The state a transformation can be in.
